% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_plot.R
\name{boc_plot}
\alias{boc_plot}
\title{Plot Bank of Canada time series}
\usage{
boc_plot(df)
}
\arguments{
\item{df}{A tibble returned by boc_series()}
}
\value{
A ggplot object
}
\description{
Plot Bank of Canada time series
}
