% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_boc_request.R
\name{boc_request}
\alias{boc_request}
\title{Internal helper to call Bank of Canada Valet API}
\usage{
boc_request(path, query = list())
}
\arguments{
\item{path}{API path after /valet/ (e.g. "lists/series/json")}

\item{query}{Named list of query parameters (e.g. list(start_date="2020-01-01"))}
}
\value{
Parsed JSON response (list)
}
\description{
Internal helper to call Bank of Canada Valet API
}
\keyword{internal}
