% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_utils.R
\name{boc_rolling_mean}
\alias{boc_rolling_mean}
\title{Rolling mean (right-aligned)}
\usage{
boc_rolling_mean(
  data,
  value_col = "value",
  window = 5,
  order_col = "date",
  group_col = "series",
  new_col = "roll_mean"
)
}
\description{
Rolling mean (right-aligned)
}
