% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_utils.R
\name{boc_summary}
\alias{boc_summary}
\title{Summary stats by series}
\usage{
boc_summary(
  data,
  value_col = "value",
  order_col = "date",
  group_col = "series"
)
}
\description{
Summary stats by series
}
