% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_utils.R
\name{boc_autocorr}
\alias{boc_autocorr}
\title{Autocorrelation up to a chosen lag (per series)}
\usage{
boc_autocorr(data, value_col = "value", lag_max = 10, group_col = "series")
}
\description{
Autocorrelation up to a chosen lag (per series)
}
