% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_utils.R
\name{boc_percent_change}
\alias{boc_percent_change}
\title{Percent change over k periods (arithmetic or log)}
\usage{
boc_percent_change(
  data,
  value_col = "value",
  periods = 1,
  type = c("arithmetic", "log"),
  order_col = "date",
  group_col = "series",
  new_col = "pct_change"
)
}
\description{
Percent change over k periods (arithmetic or log)
}
