% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_request.R
\name{boc_request}
\alias{boc_request}
\title{Internal helper to call Bank of Canada Valet API}
\usage{
boc_request(path, query = list())
}
\arguments{
\item{path}{API path after /valet/}

\item{query}{Named list of query parameters}
}
\value{
Parsed JSON response
}
\description{
Internal helper to call Bank of Canada Valet API
}
\keyword{internal}
