% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc_series.R
\name{boc_series}
\alias{boc_series}
\title{Get Bank of Canada time series data}
\usage{
boc_series(series, start_date = NULL, end_date = NULL)
}
\arguments{
\item{series}{Character vector of series IDs (e.g. "FXUSDCAD")}

\item{start_date}{Start date (YYYY-MM-DD)}

\item{end_date}{End date (YYYY-MM-DD)}
}
\value{
A tibble with date, series, and value
}
\description{
Get Bank of Canada time series data
}
